+++ ChangeLog.md
@@ -5,6 +5,8 @@ v 0.6.0.0
 the included directory, which may be installed in a platform independent way.
 * Added two more html templates, pandoc-bootstrap-KH.html and pandoc-adaptive-bootstrap-KH.html
 with pandoc-bootstrap.css and used them in some of the examples.
+* Added the default html templates into the pandoc-templates directory where Pandoc will expect
+them if the given template fails.
 
 v 0.5.0.0
 * Added plots example back since there is a version of plots on hackage with