+++ ChangeLog.md
@@ -7,6 +7,7 @@ the included directory, which may be installed in a platform independent way.
 with pandoc-bootstrap.css and used them in some of the examples.
 * Added the default html templates into the pandoc-templates directory where Pandoc will expect
 them if the given template fails.
+* Reorganized data directory with "pandoc-data" "knit-haskell-templates" and "knit-haskell-css"
 
 v 0.5.0.0
 * Added plots example back since there is a version of plots on hackage with